Move unlock_luks2_dev()
